fix: close audit P0–P2 remediations; monthly UI evals
Unblock audit:full (tar/js-yaml/brace-expansion), move UI evals to monthly,
close security footguns (IMDS deny, unsafeAllowNoAuth startup refuse, length-only
credential diagnostics, constant-time parity), and ship hygiene/structural
follow-ups from the 2026-07-21 repo-wide surface audit.
