+++ .github/pull_request_template.md
@@ -6,10 +6,21 @@
 ## Validation
 
 - [ ] `pnpm test`
+- [ ] `pnpm test:gateway` (when Durable Gateway / `products/gateway` or shared security primitives change)
 - [ ] `pnpm typecheck`
 - [ ] `pnpm lint`
 - [ ] `pnpm perf:check` (if touched runtime, projection, or performance-sensitive code)
 - [ ] `git diff --check`
+
+## Dual-channel security checklist
+
+Required when this PR touches **channel security or protocol** (signatures, SSRF/webhook URLs, auth tokens, rate limits, constant-time compares, trusted-target allowlists). See `docs/product-channel-ownership.md`.
+
+- [ ] N/A — not a channel security/protocol change
+- [ ] Reviewed **monorepo providers** (`packages/gateway-provider-*`, `packages/gateway-channel`, `apps/channel-gateway`)
+- [ ] Reviewed **Durable Gateway channels** (`products/gateway/src/channels/*` and related security)
+- [ ] Shared primitives preferred (`@open-cowork/shared`, `gateway-channel`) over copy-paste
+- [ ] Both stacks fixed **or** explicit single-stack ownership noted in Notes with follow-up
 
 ## User-visible impact
