+++ apps/desktop/src/main/cloud-workspace-adapter.ts
@@ -1,3 +1,4 @@
+import type { WorkspaceSessionPort } from './workspace-session-port.ts'
 import type {
   ArtifactIndexPayload,
   ArtifactIndexRequest,
@@ -80,27 +81,19 @@ export type CloudPromptInput = {
   attachments?: MessageAttachment[]
 }
 
-export type CloudWorkspaceSessionAdapter = {
-  policy(): Promise<WorkspacePolicy>
+/**
+ * Cloud workspace session adapter. Extends the shared {@link WorkspaceSessionPort}
+ * surface so cloud and Durable Gateway bridges share one session/workflow contract
+ * (audit 2026-07-21 P2-8). Optional methods remain cloud-only extensions.
+ */
+export type CloudWorkspaceSessionAdapter = WorkspaceSessionPort & {
   sync?(): Promise<void>
-  listSessions(): Promise<SessionInfo[]>
   createSession(input?: { projectSource?: CloudProjectSourceInput | null }): Promise<SessionInfo>
   validateProjectSource?(input: CloudProjectSourceInput): Promise<CloudProjectSourcePolicyVerdict>
   uploadProjectSnapshot?(input: CloudProjectSnapshotUploadInput): Promise<CloudProjectSnapshotUploadResult>
   importSession(input: SessionImportRequest): Promise<{ session: SessionInfo, view: SessionView }>
-  getSessionInfo(sessionId: string): Promise<SessionInfo | null>
-  getSessionView(sessionId: string): Promise<SessionView>
   promptSession(sessionId: string, input: CloudPromptInput): Promise<void>
-  abortSession(sessionId: string): Promise<void>
-  replyToQuestion?(sessionId: string, requestId: string, answers: unknown[]): Promise<void>
-  rejectQuestion?(sessionId: string, requestId: string): Promise<void>
-  respondToPermission?(sessionId: string, permissionId: string, allowed: boolean): Promise<void>
   listWorkflows?(input?: WorkflowListRequest): Promise<WorkflowListPayload>
-  getWorkflow?(workflowId: string): Promise<WorkflowDetail | null>
-  runWorkflow?(workflowId: string): Promise<WorkflowRun | null>
-  pauseWorkflow?(workflowId: string): Promise<WorkflowDetail | null>
-  resumeWorkflow?(workflowId: string): Promise<WorkflowDetail | null>
-  archiveWorkflow?(workflowId: string): Promise<WorkflowDetail | null>
   searchThreads?(query?: ThreadSearchQuery): Promise<ThreadSearchResult>
   threadFacets?(query?: ThreadSearchQuery): Promise<ThreadFacetSummary>
   listThreadTags?(): Promise<ThreadTag[]>
