+++ package.json
@@ -115,10 +115,12 @@
     "changeset:version": "changeset version",
     "build:channel-gateway": "pnpm --filter @open-cowork/channel-gateway build",
     "typecheck:channel-gateway": "pnpm --filter @open-cowork/channel-gateway typecheck",
-    "boundaries:check": "node scripts/check-product-boundaries.mjs",
+    "boundaries:check": "node scripts/check-product-boundaries.mjs && node scripts/check-no-ghost-apps-gateway.mjs",
     "smoke:gateway-standalone": "node products/gateway/scripts/standalone-smoke.mjs",
     "smoke:wiki-standalone": "node products/wiki/scripts/standalone-smoke.mjs",
-    "archive:readiness": "node scripts/check-product-archive-readiness.mjs"
+    "archive:readiness": "node scripts/check-product-archive-readiness.mjs",
+    "test:gateway": "pnpm --filter cowork-gateway test",
+    "test:gateway:typecheck": "pnpm --filter cowork-gateway typecheck"
   },
   "engines": {
     "node": ">=22.22.3",
