+++ pnpm-workspace.yaml
@@ -43,10 +43,17 @@ overrides:
   "form-data@<4.0.6": ">=4.0.6"
   "hono@<4.12.25": ">=4.12.25"
   "ip-address@<=10.1.0": ">=10.1.1"
-  "js-yaml@<4.2.0": ">=4.2.0 <5"
+  # GHSA-52CP-R559-CP3M: merge-key quadratic DoS; fixed in 4.3.0 / 3.15.0.
+  "js-yaml@<4.3.0": ">=4.3.0 <5"
+  "js-yaml@>=3.0.0 <3.15.0": ">=3.15.0 <4"
   "mermaid>uuid": ^14.0.0
   "qs@>=6.11.1 <=6.15.1": 6.15.2
-  "tar@>=7.0.0 <7.5.16": ">=7.5.16"
+  # GHSA-23HP-3JRH-7FPW (crit) + GHSA-8X88-C5MF-7J5W (high): node-tar DoS.
+  "tar@>=7.0.0 <7.5.19": ">=7.5.19"
+  # GHSA-3JXR-9VMJ-R5CP: brace-expansion exponential DoS (eslint/minimatch graph).
+  "brace-expansion@<1.1.16": ">=1.1.16 <2"
+  "brace-expansion@>=2.0.0 <2.1.2": ">=2.1.2 <3"
+  "brace-expansion@>=3.0.0 <5.0.7": ">=5.0.7"
   "tmp@<0.2.7": 0.2.7
   "undici@>=7.0.0 <7.28.0": ">=7.28.0 <8"
   "undici@<6.27.0": ^6.27.0
